refactor: SoC
