+++ client/src/components/index.ts
@@ -1,0 +1,3 @@
+export * from "./button";
+export * from "./input-group";
+export * from "./message-list";
