+++ client/src/components/input-group/index.ts
@@ -1,0 +1,1 @@
+export { InputGroup } from "./input-group";
