+++ client/src/components/message-list/index.ts
@@ -1,0 +1,1 @@
+export { MessageList } from "./message-list";
