+++ client/src/hooks/chat/index.ts
@@ -1,0 +1,1 @@
+export { useChat } from "./use-chat.hook";
