+++ client/src/types/chat/index.ts
@@ -1,0 +1,1 @@
+export type { Message } from "./message.type";
