+++ client/src/types/chat/message.type.ts
@@ -1,0 +1,5 @@
+export interface Message {
+    id: number;
+    text: string;
+    fromUser: boolean;
+}
