+++ client/tsconfig.app.json
@@ -18,10 +18,12 @@
         "strict": true,
         "noUnusedLocals": true,
         "noUnusedParameters": true,
-        "noFallthroughCasesInSwitch": true
+        "noFallthroughCasesInSwitch": true,
+
+        "baseUrl": ".",
+        "paths": {
+            "@/*": ["./src/*"]
+        }
     },
-    "include": ["src"],
-    "paths": {
-        "@/*": ["./src/*"]
-    }
+    "include": ["src/**/*"]
 }
